+++ interface/lib/interface.dart
@@ -4,4 +4,4 @@ library device_vendor_info_interface;
 export 'src/info/bios.dart';
 export 'src/info/board.dart';
 export 'src/info/system.dart';
-export 'src/loader.dart';
+export 'src/loader.dart' show DeviceVendorInfoLoader;