+++ interface/lib/mock.dart
@@ -1,0 +1,4 @@
+/// Emulation library for fetching dummy hardware information.
+library mock;
+
+export 'src/loader.dart' show MockDeviceVendorInfoLoader;