+++ interface/lib/release.dart
@@ -1,0 +1,6 @@
+/// Release version for extracting real hardware information
+/// during release.
+library release;
+
+export 'src/dictionary.dart';
+export 'src/loader.dart' show ProductiveDeviceVendorInfoLoader;