+++ interface/lib/src/definitions.dart
@@ -1,5 +1,0 @@
-import 'package:intl/intl.dart' show DateFormat;
-
-/// Definition of [DateFormat] that indicate release date
-/// of BIOS.
-final DateFormat biosDateFormat = DateFormat("MM/dd/yyyy");